numerous, poorly tracked updates to the scripts. This is mostly just a test of my new commit template
this is a midway point in this initial attempt at packaging software. I
need to do a lot of work in the basic pipeline processing to standardize
the column headings, etc.

At this point, be very careful of what columns are in the scripts. those
are likely EXTREMELY BUGGY right now. Will be fixed shortly

[Ticket: none...not using this feature yet, but will start soon]